First version detail window:
Added the first version of the movie app detail window.

Actors and movie detail components were created.

Interfaces were created to handle the information provided by the api for actor details and movie details.

The hook was created to handle the iteration with the api.

Icon libraries were added, this time we are using font-awesone as icon manager in our application.